1.15-pre6

+++ net/minecraft/nbt/ByteArrayTag.java
@@ -26,7 +26,7 @@ extends AbstractListTag<ByteTag> {
         public ByteArrayTag read(DataInput dataInput, int i, PositionTracker positionTracker) throws IOException {
             positionTracker.add(192L);
             int j = dataInput.readInt();
-            positionTracker.add(8 * j);
+            positionTracker.add(8L * (long)j);
             byte[] bs = new byte[j];
             dataInput.readFully(bs);
             return new ByteArrayTag(bs);
